+++ tests/TestCase.php
@@ -82,9 +82,9 @@ abstract class TestCase extends BaseTestCase
             {
             }
 
-            public function httpRequest(Server $server, string $method, string $url, ?array $json = null, int $timeoutSeconds = 15): array
+            public function httpRequest(Server $server, string $method, string $url, ?array $json = null, int $timeoutSeconds = 15, array $headers = []): array
             {
-                $request = Http::timeout($timeoutSeconds)->acceptJson();
+                $request = Http::timeout($timeoutSeconds)->acceptJson()->withHeaders($headers);
 
                 if ($json !== null) {
                     $request = $request->asJson();